Refactoring Exceptions Related to Users
