--- src/main/java/com/unsa/suppliers/domain/exceptions/UserNotFoundException.java
@@ -1,9 +1,0 @@
-package com.unsa.suppliers.domain.exceptions;
-
-import static com.unsa.suppliers.domain.exceptions.UserExceptionMessages.NOT_FOUND;
-
-public class UserNotFoundException extends Exception {
-    public UserNotFoundException() {
-        super(NOT_FOUND);
-    }
-}
