+++ src/main/java/com/unsa/suppliers/domain/exceptions/users/UserNotFoundException.java
@@ -1,0 +1,9 @@
+package com.unsa.suppliers.domain.exceptions.users;
+
+import static com.unsa.suppliers.domain.exceptions.users.UserExceptionMessages.NOT_FOUND;
+
+public class UserNotFoundException extends Exception {
+    public UserNotFoundException() {
+        super(NOT_FOUND);
+    }
+}
